requirements.txt
groq>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
streamlit>= 1.54.0
